--- node-js-getting-started/cloud.js
@@ -1,8 +1,0 @@
-var AV = require('leanengine');
-
-/**
- * 一个简单的云代码方法
- */
-AV.Cloud.define('hello', function(request, response) {
-  response.success('Hello world!');
-});
