--- node-js-getting-started/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "node-js-getting-started",
-  "version": "1.0.0",
-  "private": true,
-  "description": "A sample Node.js app using Express",
-  "main": "server.js",
-  "scripts": {
-    "start": "node server.js",
-    "dev": "nodemon server.js"
-  },
-  "dependencies": {
-    "body-parser": "1.12.3",
-    "connect-timeout": "^1.7.0",
-    "cookie-parser": "^1.3.5",
-    "ejs": "2.3.1",
-    "express": "4.12.3",
-    "leanengine": "^1.2.3"
-  },
-  "devDependencies": {
-    "nodemon": "^1.11.0"
-  },
-  "engines": {
-    "node": "6.x"
-  }
-}
