--- node-js-getting-started/views/error.ejs
@@ -1,12 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-  <head>
-    <title>Error</title>
-    <link rel="stylesheet" href="/stylesheets/style.css">
-  </head>
-  <body>
-    <h1><%= message %></h1>
-    <h2><%= error.status %></h2>
-    <pre><%= error.stack %></pre>
-  </body>
-</html>
