+++ views/index.ejs
@@ -1,0 +1,13 @@
+<!DOCTYPE HTML>
+<html>
+  <head>
+    <title>LeanEngine</title>
+    <link rel="stylesheet" href="/stylesheets/style.css">
+  </head>
+  <body>
+    <h1>LeanEngine</h1>
+    <p>这是 LeanEngine 的示例应用</p>
+    <p>当前时间：<%= currentTime %></p>
+    <p><a href="/todos">一个简单的「TODO 列表」示例</a></p>
+  </body>
+</html>
