+++ views/todos.ejs
@@ -1,0 +1,19 @@
+<!DOCTYPE HTML>
+<html>
+  <head>
+    <title>Todo</title>
+    <link rel="stylesheet" href="/stylesheets/style.css">
+  </head>
+  <body>
+    <h1><%= title %></h1>
+    <form action="/todos" method="POST">
+      <input type="text" name="content" />
+      <input type="submit" value="新增" />
+    </form>
+    <ul>
+      <% for(var i=0; i<todos.length; i++) {%>
+      <li><%= todos[i].get('content') %></li>
+      <% } %>
+    </ul>
+  </body>
+</html>
